Adding the installation mode
The installation mode is a function to set a mersenne order after the installation of program.
The submitted order is tested by thye program, saved in the hotsave file, and the program is
closed. At the next start, the program starts prospecting at a valid mersenne order...

Man is up to date